refactoring. missing changes to framework pages
