+++ pages/Frameworks.py
@@ -3,15 +3,15 @@
 import os
 import streamlit as st
 from dotenv import load_dotenv
-from utilities.utils_openai import generate_advanced_query
-from pages.frameworks.COG import cog_analysis
-from pages.frameworks.SWOT import swot_page
-from pages.frameworks.ACH import ach_page
-from pages.frameworks.DeceptionDetection import deception_detection
-from pages.frameworks.DIME import dime_page
+from utilities.gpt import generate_advanced_query
+from frameworks.cog import cog_analysis
+from frameworks.swot import swot_page
+from frameworks.ach import ach_page
+from frameworks.deception_detection import deception_detection
+from frameworks.dime import dime_page
 from sidebar_menu import sidebar_menu
-from pages.frameworks.PMESII_PT import pmesi_page
-from pages.frameworks.DOTMLPF import dotmlpf_page
+from frameworks.pmesii_pt import pmesii_pt_page
+from frameworks.dotmlpf import dotmlpf_page
 
 load_dotenv()
