+++ tests/standalone_locations_to_kml.py
